requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
billiard==3.6.4.0
blessed==1.20.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-environ==0.10.0
django-jazzmin==2.6.0
django-picklefield==3.1
h11==0.14.0
idna==3.4
kombu==5.2.4
outcome==1.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pydub==0.25.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==3.5.3
requests==2.28.2
schedule==1.1.0
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
